+++ packages/react/src/composite/Modal/src/components/ModalHeader/ModalHeader.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import { Title } from "@nimbus-ds/title";
-import { Box } from "@nimbus-ds/box";
 import { modal } from "@nimbus-ds/styles";
 
 import { ModalHeaderProps } from "./modalHeader.types";
@@ -16,14 +15,14 @@ const ModalHeader: React.FC<ModalHeaderProps> = ({
 }) => (
   <div className={modal.subComponents.header.sprinkle({ padding })} {...rest}>
     {(title || tag) && (
-      <Box display="flex" alignItems="center" gap="2">
+      <div className={modal.subComponents.header.content}>
         {title && (
           <Title data-testid="header-title" as="h4">
             {title}
           </Title>
         )}
         {tag}
-      </Box>
+      </div>
     )}
     {children}
   </div>